+++ problem/induced_path.go
@@ -4,6 +4,7 @@ import (
 	"slices"
 	"strings"
 
+	"github.com/roidaradal/fn/dict"
 	"github.com/roidaradal/fn/list"
 	"github.com/roidaradal/opt/data"
 	"github.com/roidaradal/opt/discrete"
@@ -16,6 +17,10 @@ func NewInducedPath(variant string, n int) *discrete.Problem {
 	switch variant {
 	case "basic":
 		return maxInducedPath(name)
+	case "weighted":
+		return maxWeightedInducedPath(name)
+	case "k":
+		return kInducedPath(name)
 	default:
 		return nil
 	}
@@ -78,10 +83,53 @@ func maxInducedPath(name string) *discrete.Problem {
 
 	p.ObjectiveFn = func(solution *discrete.Solution) discrete.Score {
 		// Number of points in path (not assigned -1)
-		points := list.Filter(solution.Values(), func(value discrete.Value) bool {
-			return value >= 0
-		})
+		points := inducedPath(solution)
 		return discrete.Score(len(points))
 	}
 	return p
+}
+
+// Max Weighted Induced Path
+func maxWeightedInducedPath(name string) *discrete.Problem {
+	p, cfg := newInducedPathProblem(name)
+	if p == nil || cfg == nil {
+		return nil
+	}
+	if len(cfg.VertexWeight) != len(cfg.Vertices) {
+		return nil
+	}
+
+	p.ObjectiveFn = func(solution *discrete.Solution) discrete.Score {
+		variables := list.Map(inducedPath(solution), func(e dict.Entry[discrete.Variable, discrete.Value]) discrete.Variable {
+			return e.Key
+		})
+		return list.Sum(list.MapList(variables, cfg.VertexWeight))
+	}
+	return p
+}
+
+// K-Induced Path
+func kInducedPath(name string) *discrete.Problem {
+	p, cfg := newInducedPathProblem(name)
+	if p == nil || cfg == nil {
+		return nil
+	}
+	if cfg.K == 0 {
+		return nil
+	}
+
+	p.Goal = discrete.Satisfy
+	p.AddUniversalConstraint(func(solution *discrete.Solution) bool {
+		// Number of points in path (not assigned -1)
+		points := inducedPath(solution)
+		return len(points) == cfg.K
+	})
+	return p
+}
+
+// Common step for getting induced path points from the solution
+func inducedPath(solution *discrete.Solution) []dict.Entry[discrete.Variable, discrete.Value] {
+	return list.Filter(solution.Entries(), func(e dict.Entry[discrete.Variable, discrete.Value]) bool {
+		return e.Value >= 0
+	})
 }